[MIGR] mrp_calendar_view: migrate it to v10.0 and rename it to mrp_repair_calendar_view
