--- mrp_calendar_view/README.rst
@@ -1,29 +1,0 @@
-.. image:: https://img.shields.io/badge/licence-AGPL--3-blue.svg
-    :alt: License
-
-MRP Calendar View
-=================
-
-This Module adds a calendar view for the repair orders
-
-Credits
-=======
-
-Contributors
-------------
-
-* Nicola Malcontenti <nicola.malcontenti@agilebg.com>
-* Odoo Community Association
- 
-Maintainer
-----------
-
-.. image:: http://odoo-community.org/logo.png
-   :alt: Odoo Community Association
-   :target: http://odoo-community.org
-
-This module is maintained by the OCA.
-
-OCA, or the Odoo Community Association, is a nonprofit organization whose mission is to support the collaborative development of Odoo features and promote its widespread use.
-
-To contribute to this module, please visit http://odoo-community.org.
