+++ mrp_repair_calendar_view/__init__.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+# Copyright 2016  Agile Business Group
+# License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl).
+
+
+from . import models
