+++ mrp_repair_calendar_view/__manifest__.py
@@ -1,0 +1,16 @@
+# -*- coding: utf-8 -*-
+# Copyright 2016  Agile Business Group
+# License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl).
+
+
+{
+    'name': 'MRP Repair Calendar View',
+    'version': '10.0.1.0.0',
+    'category': 'Manufacturing',
+    'author': "Agile Business Group, Odoo Community Association (OCA)",
+    'website': 'http://www.agilebg.com',
+    'license': 'LGPL-3',
+    'depends': ['mrp_repair'],
+    'data': ['views/mrp_repair_view.xml'],
+    'installable': True,
+}
