+++ mrp_repair_calendar_view/i18n/it.po
@@ -1,0 +1,47 @@
+# Translation of Odoo Server.
+# This file contains the translation of the following modules:
+#	* mrp_repair_calendar_view
+#
+msgid ""
+msgstr ""
+"Project-Id-Version: Odoo Server 10.0\n"
+"Report-Msgid-Bugs-To: \n"
+"POT-Creation-Date: 2017-04-07 08:27+0000\n"
+"PO-Revision-Date: 2017-04-07 08:27+0000\n"
+"Last-Translator: <>\n"
+"Language-Team: \n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: \n"
+"Plural-Forms: \n"
+
+#. module: mrp_repair_calendar_view
+#: model:ir.model.fields,help:mrp_repair_calendar_view.field_mrp_repair_date_repair
+msgid "Date of the repair, this field and user_id defines the calendar"
+msgstr "Data della riparazione, questo campo e user_id definiscono il calendario"
+
+#. module: mrp_repair_calendar_view
+#: model:ir.model.fields,help:mrp_repair_calendar_view.field_mrp_repair_user_id
+msgid "Person in charge for the repair"
+msgstr "Persona responsabile della riparazione"
+
+#. module: mrp_repair_calendar_view
+#: model:ir.model.fields,field_description:mrp_repair_calendar_view.field_mrp_repair_date_repair
+msgid "Repair Date"
+msgstr "Data riparazione"
+
+#. module: mrp_repair_calendar_view
+#: model:ir.model,name:mrp_repair_calendar_view.model_mrp_repair
+msgid "Repair Order"
+msgstr "Ordine di riparazione"
+
+#. module: mrp_repair_calendar_view
+#: model:ir.ui.view,arch_db:mrp_repair_calendar_view.view_repair_order_calendar
+msgid "Repair Orders Calendar"
+msgstr "Calendario ordini di riparazione"
+
+#. module: mrp_repair_calendar_view
+#: model:ir.model.fields,field_description:mrp_repair_calendar_view.field_mrp_repair_user_id
+msgid "Responsible"
+msgstr "Responsabile"
+
