+++ mrp_repair_calendar_view/views/mrp_repair_view.xml
@@ -1,7 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<openerp>
-    <data>
-
+<odoo>
     <record id="view_repair_order_calendar_form" model="ir.ui.view">
         <field name="name">mrp.repair.calendar.form</field>
         <field name="model">mrp.repair</field>
@@ -31,5 +29,4 @@
         <field name="view_mode">calendar,tree,form</field>
     </record>
 
-    </data>
-</openerp>
+</odoo>
